refactor: separação de responsabilidade, atualizar estoque, realizar venda

+++ src/main/java/com/tauan_estoque_venda/dtos/ItemPedidoDTO.java
@@ -1,5 +1,5 @@
 package com.tauan_estoque_venda.dtos;
 
-public record ItemPedidoDTO(Long produtoId, Integer quantidade) {
+public record ItemPedidoDTO(Integer produtoId, Integer quantidade) {
 
 }
